Finally getting information from loaded save file.
The entire saving and loading system had to be redesigned.
We cannot serialize pointers and expect they'll be pointing to the memory that magically holds previous content on the heap.
It's best to include an 'intrusive' version of serializing data - passing a friend access class to the struct/class and writing serializers directly in class.